+++ zebra-state/src/service/finalized_state/zebra_db/block.rs
@@ -57,6 +57,8 @@ use crate::{
 #[cfg(feature = "indexer")]
 use crate::request::Spend;
 
+mod startup_audit;
+
 #[cfg(test)]
 mod tests;